feat(api): complete P45A module split (#775)
- reduce main.py to composition root only
- extract watchlists + analysis routers
- move business logic to analysis_service and inspection_service
- introduce composition_runtime_service for wiring/runtime concerns
- align architecture documentation

No API behavior changes. Full pytest suite passing (729 tests).
